requirements.txt
django~=2.2
Pillow==7.0.0
psycopg2-binary==2.8.4 
django-leaflet==0.26.0
black
django-widget-tweaks==1.4.8
geopy==1.21.0
Unidecode==1.1.1
djangorestframework==3.11.0
django-filter==2.2.0
djangorestframework-gis==0.15
django-silk==4.0.1
ua-parser~=0.10
django-environ==0.4.5
graphene_django==2.9.0
django-graphql-geojson==0.1.4
